added parameters, resolved problems
