+++ shift_detector/checks/simple_check.py
@@ -11,12 +11,28 @@ from shift_detector.utils.column_management import ColumnType
 
 class SimpleCheck(Check):
 
-    def __init__(self):
+    def __init__(self, categorical_threshold=0.05, mean_threshold=0.1, median_threshold=0.1, min_threshold=0.15,
+                 max_threshold=0.15, quartile_1_threshold=0.15, quartile_3_threshold=0.15, uniqueness_threshold=0.1,
+                 num_distinct_threshold=0.1, completeness_threshold=0.1, std_threshold=0.1):
+
+        threshold_names_values = {'mean': mean_threshold, 'median': median_threshold, 'min': min_threshold,
+                                  'max': max_threshold, 'quartile_1': quartile_1_threshold,
+                                  'quartile_3': quartile_3_threshold, 'uniqueness': uniqueness_threshold,
+                                  'num_distinct': num_distinct_threshold, 'std': std_threshold,
+                                  'completeness': completeness_threshold}
+
+        if categorical_threshold < 0 or categorical_threshold > 1:
+            raise ValueError('The categorical threshold of {} is not correct. It has be between the values of '
+                             '0 and 1.'.format(categorical_threshold))
+
+        for t_name, t_value in threshold_names_values.items():
+            if t_value < 0 or t_value > 1:
+                raise ValueError('The {}_threshold of {} is not correct. It has be between the values of 0 and 1'
+                                 .format(t_name, t_value))
+
         self.data = None
-        self.categorical_threshold = 0.05
-        self.metrics_thresholds_percentage = {'mean': 10, 'median': 10, 'min': 15, 'max': 15, 'quartile_1': 15,
-                                              'quartile_3': 15, 'uniqueness': 10, 'num_distinct': 10,
-                                              'completeness': 10, 'std': 10}
+        self.categorical_threshold = categorical_threshold
+        self.metrics_thresholds_percentage = threshold_names_values
 
     def run(self, store):
         logger.info("Execute Simple Check")
@@ -33,12 +49,12 @@ class SimpleCheck(Check):
 
         if metric_in_df1 == 0 and metric_in_df2 == 0:
             return 0
-        # TODO: think about comparison if base value is 0
         if metric_in_df1 == 0:
-            logger.warning('column {} \t \t {}: no comparison of distance possible, division by zero'.format(column, metric_name))
+            logger.warning('column {} \t \t {}: no comparison of distance possible, division by zero'
+                           .format(column, metric_name))
             return 0
 
-        relative_difference = (metric_in_df2 / metric_in_df1 - 1) * 100
+        relative_difference = (metric_in_df2 / metric_in_df1 - 1)
         if metric_name in ['uniqueness', 'completeness', 'completeness']:
             relative_difference = metric_in_df2 - metric_in_df1
 
@@ -65,7 +81,8 @@ class SimpleCheck(Check):
                 diff = self.relative_metric_difference(column_name, metric)
                 diff = round(diff, 2)
 
-                if abs(diff) > self.metrics_thresholds_percentage[metric]:
+                # -.01 for rounding errors
+                if abs(diff) > self.metrics_thresholds_percentage[metric] - .01:
                     shifted_columns.add(column_name)
                     explanation[column_name] += "Metric: {} with Diff: {}\n".format(metric,
                                                                                     self.difference_to_string(diff))
@@ -117,12 +134,13 @@ class SimpleReport(Report):
             num_columns = len(list(df1.columns))
             for num, column in enumerate(list(df1.columns)):
                 a, b = df1[column], df2[column]
-                ax = f.add_subplot(1, num_columns, num+1)
+                ax = f.add_subplot(1, num_columns, num + 1)
 
                 ax.boxplot([a, b])
                 ax.set_title(column)
 
             plt.show()
+
         return custom_plot
 
     @staticmethod
@@ -135,7 +153,7 @@ class SimpleReport(Report):
             for i, plot_info in enumerate(list(plot_infos)):
                 bars1, bars2, attribute_names, column_name = plot_info[0], plot_info[1], plot_info[2], plot_info[3]
 
-                subplot = f.add_subplot(1, num_columns, i+1)
+                subplot = f.add_subplot(1, num_columns, i + 1)
 
                 bar_width = 0.25
                 r1 = np.arange(len(bars1))
@@ -146,11 +164,10 @@ class SimpleReport(Report):
 
                 subplot.title.set_text(column_name)
                 subplot.set_xlabel('attribute-values', fontweight='bold')
-                subplot.set_xticks(np.arange(len(attribute_names))+bar_width/2)
+                subplot.set_xticks(np.arange(len(attribute_names)) + bar_width / 2)
                 subplot.set_xticklabels(attribute_names)
                 subplot.legend()
 
             f.show()
 
         return custom_plot
-
